+++ crates/inertia-embed/src/lib.rs
@@ -6,5 +6,5 @@ mod cache;
 mod frontend;
 mod request;
 
-pub use frontend::{EmbeddedAsset, EmbeddedFrontend};
+pub use frontend::{EmbeddedAsset, EmbeddedFrontend, EmbeddedStorage};
 pub use inertia_embed_macros::embed_frontend;